requirements.txt
aiogram==2.21
python-dotenv==1.0.0
requests==2.32.3